code/flask-server/requirements.txt
asgiref==3.4.1
boto3==1.18.13
botocore==1.21.13
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
Flask==2.0.1
idna==3.2
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==2.0.1
